Remoção do README e Script
- Remoção de README não utilizado
- Script de integração da viaCep realocado para ser carregado no layout
